Adding custom validations
