+++ app/assets/stylesheets/users.css.scss
@@ -6,77 +6,77 @@
 //  width: 280px;
 //  height:70px;
 //
-//padding:10px;     
+//padding:10px;
 //box-shadow:3px 3px 7px #0f0f0f;
 //
-////border: 1px solid #4c4c4c;   
+////border: 1px solid #4c4c4c;
 //
 //
-//background: #7abcff; 
+//background: #7abcff;
 //
 //
 // -moz-border-radius: 5px;
 //border-radius: 5px;
 //
 //
-//float:left;  
-//margin-right:50px;    
+//float:left;
+//margin-right:50px;
 
 background: url("small_user.png") no-repeat scroll left center #60abf8 !important;
-	
-}   
+
+}
 
 //#user ul {
-//	float:right;
-//	width:150px; 
-//}     
+//  float:right;
+//  width:150px;
+//}
 
-  
+
 
 #user_nodes ul{
-	margin-top:5px;
+  margin-top:5px;
 }
-#node_list li {	
-	 width: 170px;
-	 box-shadow:1px 1px 1px #0f0f0f;
-	 -moz-border-radius: 10px;
-	 border-radius: 10px;
-	 margin-bottom:10px; 
-	 background:#60ABF8;
+#node_list li {
+   width: 170px;
+   box-shadow:1px 1px 1px #0f0f0f;
+   -moz-border-radius: 10px;
+   border-radius: 10px;
+   margin-bottom:10px;
+   background:#60ABF8;
 }
 
 
-#node_list li.on { 
-     background: url("mini_on.png") no-repeat scroll 8px center #60ABF8 !important; 	
-	 }                          
+#node_list li.on {
+     background: url("mini_on.png") no-repeat scroll 8px center #60ABF8 !important;
+   }
 
-#node_list li.off { 
-	background: url("mini_off.png") no-repeat scroll 8px center #60ABF8 !important;  
-	}
+#node_list li.off {
+  background: url("mini_off.png") no-repeat scroll 8px center #60ABF8 !important;
+  }
 
 #node_list li.error {
-	background: url("mini_error.png") no-repeat scroll 8px center #60ABF8 !important;  
-}   
+  background: url("mini_error.png") no-repeat scroll 8px center #60ABF8 !important;
+}
 
 #node_list li a {
-	margin-left:40px;
-}   
+  margin-left:40px;
+}
 
-#user_nodes { 
-	clear:both;
-	width:280px;
-	background:#7abcff;
-	margin-left:10px;       
-	box-shadow:1px 1px 3px #0f0f0f;  
-	padding-top:10px; 
-	padding-bottom:5px;
-}     
+#user_nodes {
+  clear:both;
+  width:280px;
+  background:#7abcff;
+  margin-left:10px;
+  box-shadow:1px 1px 3px #0f0f0f;
+  padding-top:10px;
+  padding-bottom:5px;
+}
 
 
 //form p {
 //  padding: 7px 0;
 //  font-size: 12px;
-//}    
+//}
 
 form a {
   color: #000;
@@ -89,28 +89,30 @@ form p label {
   vertical-align: top;
 }
 
-form p input[type=text], 
-form p input[type=url], 
+form p input[type=text],
+form p input[type=url],
 form p input[type=number] {
   border: 1px solid #809db9;
   font-size: 12px;
   padding: 2px 4px;
   vertical-align: top;
   width: 130px;
-}    
+}
 
 
-form p input.error[type=text], 
-form p input.error[type=url], 
+form p input.error[type=text],
+form p input.error[type=url],
 form p input.error[type=number] {
   border: 1px solid red !important;
   background:none !important;
-}  
+}
 
-label.error { 
-	color:red;
-	background:none !important;
-	
+label.error {
+  color:red;
+  background:none !important;
+  padding-left: 10px;
+  width: 250px !important;
+
 }
 
 form p input[type=radio] {
@@ -128,7 +130,7 @@ form p input[type=button] {
 
 form select {
   vertical-align: top;
-}  
+}
 
 label.required {
 }
@@ -137,16 +139,16 @@ p.hint {
   font-size: 10px !important;
   font-style: italic;
   margin-top: -10px;
-}  
+}
 
 p.long input,
 p.long select{
   width:420px !important;
-}   
+}
 
 p.giant input,
 p.giant select{
-   width:700px !important;    
+   width:700px !important;
 }
 
 p.medium input,
